fixed general clippy warnings

+++ state-chain/pallets/cf-emissions/src/lib.rs
@@ -286,7 +286,7 @@ pub mod pallet {
 		fn build(&self) {
 			ValidatorEmissionInflation::<T>::put(self.validator_emission_inflation);
 			BackupValidatorEmissionInflation::<T>::put(self.backup_validator_emission_inflation);
-			MintInterval::<T>::put(T::BlockNumber::from(100 as u32));
+			MintInterval::<T>::put(T::BlockNumber::from(100_u32));
 		}
 	}
 }
